Rename project to oll-sc
