+++ README.md
@@ -1,3 +1,3 @@
-# TAF Smard Card support
+# Smart Card support
 
-This library is a wrapper around [PyKCS11](https://github.com/LudovicRousseau/PyKCS11) with included prebuild `openSC` library which allows us to sign data with smart cards (YubiKey).
+This library is a wrapper around [PyKCS11](https://github.com/LudovicRousseau/PyKCS11) with included prebuild `openSC` libraries which allows signing data with smart cards without any additional pre-installed software.
