+++ oll_sc/exceptions.py
