+++ oll_sc/opensc/opensc-pkcs11-Darwin-64bit.so
